pointing_model/requirements.txt
numpy
pandas
jupyter
matplotlib
sklearn
scipy
tensorflow
keras
seaborn
virtualenv
progressbar2
scikit-learn
# torch===1.3.1
# torchvisiontorchvision===0.4.2
featuretools